refactor: criada categoria patterns no src
componentes [About, Banner, ContactForm, Footer, Header, Projects] movidos para a categoria
patterns.
