+++ pages/index.js
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
-import Box from '../src/components/foundation/layout/Box';
-import Banner from '../src/components/commons/Banner';
-import Header from '../src/components/commons/Header';
-import Projects from '../src/components/commons/Projects';
-import Footer from '../src/components/commons/Footer';
-import theme from '../src/theme/index';
-import About from '../src/components/commons/About';
-import BackLayer from '../src/components/foundation/layout/BackLayer';
 import Modal from '../src/components/commons/Modal';
-import ContactForm from '../src/components/commons/ContactForm';
+import BackLayer from '../src/components/foundation/layout/BackLayer';
+import Box from '../src/components/foundation/layout/Box';
+import About from '../src/components/patterns/About';
+import Banner from '../src/components/patterns/Banner';
+import ContactForm from '../src/components/patterns/ContactForm';
+import Footer from '../src/components/patterns/Footer';
+import Header from '../src/components/patterns/Header';
+import Projects from '../src/components/patterns/Projects';
+import theme from '../src/theme/index';
 
 export default function Home() {
   const [isModalOpen, setIsModalOpen] = useState(false);
