+++ src/components/commons/Modal/index.js
@@ -2,12 +2,15 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import ModalWrapper, { LockScroll } from './styles';
 
+const modalId = 'ModalWrapper';
+
 export default function Modal({ isModalOpen, onClose, children }) {
   if (isModalOpen) {
     return (
       <ModalWrapper
+        id={modalId}
         onClick={(event) => {
-          if (event.target.tagName === 'SECTION') onClose();
+          if (event.target.id === modalId) onClose();
         }}
       >
         {isModalOpen && <LockScroll />}
