+++ src/components/patterns/About/styles.js
