+++ src/components/patterns/Banner/BannerTitle/index.js
