+++ src/components/patterns/Banner/index.js
