+++ src/components/patterns/Banner/styles.js
