+++ src/components/patterns/Footer/index.js
