+++ src/components/patterns/Footer/styles.js
